+++ src/index.ts
@@ -1,1 +1,3 @@
 export * from "./MongoDBDriver";
+export * from "./MongoDBModel";
+export * from "./MongoDBTimeSeriesModel";